requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autopep8==2.3.2
click==8.1.8
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
psycopg2-binary>=2.9.5
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0

